moved bookmark dropdown, added info page
moved bookmark dropdown to bookmark page because it randomly breaks the Jumpdrive page and severely limits the #of bookmarks. probably due to the massive digiline message that needs to be sent for this, in combination with some unknown server events/lag.
Added info page start with two buttons. output to events page for now.